+++ skore/src/skore/_sklearn/_checks/__init__.py
@@ -1,4 +1,4 @@
+from skore._sklearn._checks._utils import CheckNotApplicable
 from skore._sklearn._checks.base import Check, ChecksSummaryDisplay
-from skore._sklearn._checks.utils import CheckNotApplicable
 
 __all__ = ["Check", "ChecksSummaryDisplay", "CheckNotApplicable"]
